refactor code for ambient rewards timer
