+++ app/lp/utils.ts
@@ -3,7 +3,10 @@ import { AmbientTransactionParams } from "@/transactions/pairs/ambient";
 import useCantoSigner from "@/hooks/helpers/useCantoSigner";
 import useLP from "@/hooks/pairs/lpCombo/useLP";
 import { useState } from "react";
-
+import { useBlockNumber } from "wagmi";
+import { useEffect } from "react";
+import { CANTO_MAINNET_EVM } from "@/config/networks";
+import { TimeDisplayValues } from "@/hooks/pairs/newAmbient/interfaces/timeDisplay";
 export default function usePool() {
   const { txStore, signer, chainId } = useCantoSigner();
   const connectedEthAccount = signer?.account.address ?? "";
@@ -69,6 +72,58 @@ export default function usePool() {
     });
   }
 
+  /** AMBIENT REWARDS TIMER */
+
+  let prevBlockNumber = BigInt(7844908); //need to update after provided
+  const blocksInEpoch = BigInt(104272);
+  const blockDuration = 5.8;
+  let remBlocksInEpoch = BigInt(104272);
+  const { data: blockNumber } = useBlockNumber({
+    chainId: CANTO_MAINNET_EVM.chainId,
+    watch: true,
+  });
+
+  const UserAmbientRewardsTimer = (blockNumber: bigint | undefined) => {
+    let remTime = 0n;
+    if (blockNumber) {
+      const noOfWeeksToBeAdded =
+        (blockNumber - prevBlockNumber) / blocksInEpoch;
+      prevBlockNumber = prevBlockNumber + noOfWeeksToBeAdded * blocksInEpoch;
+      remBlocksInEpoch = prevBlockNumber + blocksInEpoch - blockNumber;
+      remTime = remBlocksInEpoch * BigInt(blockDuration * 1000);
+    }
+    return remTime;
+  };
+  const getTimerObj = (remTime: bigint): string => {
+    const stateObj: TimeDisplayValues = {
+      days: Number(remTime / BigInt(1000 * 60 * 60 * 24)),
+      hours: Number(
+        (remTime % BigInt(1000 * 60 * 60 * 24)) / BigInt(1000 * 60 * 60)
+      ),
+      minutes: Number((remTime % BigInt(1000 * 60 * 60)) / BigInt(1000 * 60)),
+      seconds: Number((remTime % BigInt(1000 * 60)) / BigInt(1000)),
+    };
+    return `${stateObj.days} : ${stateObj.hours} : ${stateObj.minutes} : ${stateObj.seconds}`;
+  };
+  const [ambientRewardsTimer, setAmbientRewardsTimer] = useState(
+    getTimerObj(0n)
+  );
+  useEffect(() => {
+    let remTime = remBlocksInEpoch * BigInt(blockDuration * 1000);
+    if (!blockNumber) {
+      setAmbientRewardsTimer("Loading...");
+      return;
+    }
+    remTime = UserAmbientRewardsTimer(blockNumber);
+    setInterval(() => {
+      if (remTime === 0n) {
+        remTime = UserAmbientRewardsTimer(blockNumber);
+      }
+      remTime = remTime - 1000n;
+      setAmbientRewardsTimer(getTimerObj(remTime));
+    }, 1000);
+  }, [blockNumber != undefined]);
+
   /** REWARDS */
 
   function sendClaimRewardsFlow() {
@@ -101,5 +156,6 @@ export default function usePool() {
     sendAmbientTxFlow,
     sendClaimRewardsFlow,
     pairNames,
+    ambientRewardsTimer,
   };
 }
